Fix #2 #8 finalizando docker de teste e componentizando
