+++ Client/package.json
@@ -2,16 +2,19 @@
   "name": "node-app",
   "version": "1.0.0",
   "description": "",
-  "main": "client_hospital.js",
+  "main": "app.js",
   "scripts": {
-    "start": "nodemon client_hospital.js",
-    "test": "./node_modules/.bin/mocha --reporter spec"
+    "start": "nodemon sector.js",
+    "test": "nyc --reporter=html --reporter=text mocha",
+    "coverage": "npm run cover && cat ./coverage/lcov.info | ./node_modules/coveralls/bin/coveralls.js",
+    "cover": "istanbul cover _mocha"
   },
   "author": "",
   "license": "MIT",
   "dependencies": {
-    "bcrypt": "^1.0.3",
+    "bcrypt": "2.0.1",
     "body-parser": "1.15.2",
+    "chai": "^4.1.2",
     "express": "^4.14.0",
     "jsonwebtoken": "^8.2.0",
     "mongoose": "^4.6.1",
@@ -25,17 +28,11 @@
     "seneca-web-adapter-express": "^1.1.1"
   },
   "devDependencies": {
-    "express": "^4.16.3",
-    "express-session": "^1.15.6",
+    "coveralls": "^3.0.1",
     "istanbul": "^0.4.5",
-    "jest": "^22.4.3",
-    "jsonwebtoken": "^8.2.1",
     "mocha": "^5.0.5",
-    "passport": "^0.4.0",
-    "passport-jwt": "^4.0.0",
-    "seneca": "^3.4.3",
-    "seneca-web": "^2.2.0",
-    "seneca-web-adapter-express": "^1.1.1",
+    "mocha-lcov-reporter": "^1.3.0",
+    "nyc": "^11.7.3",
     "should": "^13.2.1",
     "supertest": "^3.0.0"
   }
