+++ utils/script_coveralls.sh
@@ -1,0 +1,3 @@
+#!/bin/bash
+repo=$(git rev-parse --show-toplevel)
+echo "COVERALLS_REPO_TOKEN=${COVERALLS_REPO_TOKEN}" > ${repo}/Docker/Test/env/coveralls.env && exit 0
